Refactor layout and navigation components for admin only pages
